Initial commit to git
I decided to try out git as high profile projects (the kernel, GNOME)
were using it and I thought it was worth trying out. I was only using
bzr because of its tight integration into Launchpad, which might remain
a factor. On the other hand, I might look into some sort of interface
between the two (if git turns out to be good).