Changed check_args to use bit fields
Previously, the check_args function in suvatlib/suvatext.c used an enum
constant to find out which variables it knew. This ended up being a bit
messy for doing things like checking if we knew the displacement, so I
changed it to use an anonymous struct with bitfields instead (lifted
straight of K&R2). This makes it much nicer for things like:
if (known.s_known) {
    /* ... */
}

All the individual functions now assign to attributes of the global
variable known, whose fields are all reset to 0 by check_args once it's
finished with it.