Migrated from cdbs to debhelper
I decided that for 0.2, it would be a good idea to use the finer control
that a pure debhelper package can offer. I started that migration here,
but I will probably have completely ruined everything. This is untested,
as the package building infrastructure has not been installed on this
machine.